+++ src/lib.rs
@@ -4,5 +4,5 @@ mod extractor;
 mod scorer;
 
 pub use error::ReadabilityError;
-pub use extractor::{extract, extract_with_scorer, Product};
-pub use scorer::Scorer;
+pub use extractor::{extract, ExtractOptions, ParseOptions, Product};
+pub use scorer::{Scorer, ScorerOptions};